修复CqWsContext中Task.Run没有处理异常。修正go-cqhttp数据类型。完善DummyClient。

+++ FsCqHttp/DateType/Event/MesssageEvent.fs
@@ -68,7 +68,7 @@ type MessageEvent =
       [<JsonProperty("sub_type")>]
       SubType : string
       [<JsonProperty("message_id")>]
-      MessageId : int32
+      MessageId : int64
       [<JsonProperty("user_id")>]
       UserId : uint64
       [<JsonProperty("message")>]
